Refactor managementmodel sub-package to management
