+++ src/main/java/seedu/address/MainApp.java
@@ -19,8 +19,8 @@ import seedu.address.logic.LogicManager;
 import seedu.address.model.Lessons;
 import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.modelmanager.managementmodel.ManagementModel;
-import seedu.address.model.modelmanager.managementmodel.ManagementModelManager;
+import seedu.address.model.modelmanager.management.ManagementModel;
+import seedu.address.model.modelmanager.management.ManagementModelManager;
 import seedu.address.model.modelmanager.quizmodel.QuizModel;
 import seedu.address.model.modelmanager.quizmodel.QuizModelManager;
 import seedu.address.storage.CsvLessonImportExport;
