+++ src/main/java/seedu/address/logic/LogicManager.java
@@ -13,7 +13,7 @@ import seedu.address.logic.commands.quiz.QuizStartCommand;
 import seedu.address.logic.parser.ManagementModeParser;
 import seedu.address.logic.parser.QuizModeParser;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.modelmanager.managementmodel.ManagementModel;
+import seedu.address.model.modelmanager.management.ManagementModel;
 import seedu.address.model.modelmanager.quizmodel.QuizModel;
 
 /**
