+++ src/main/java/seedu/address/model/modelmanager/quizmodel/QuizModelManager.java
@@ -8,7 +8,7 @@ import java.util.logging.Logger;
 import seedu.address.commons.core.LogsCenter;
 
 /**
- * Represents the in-memory managementmodel of quiz data
+ * Represents the in-memory management of quiz data
  */
 public class QuizModelManager implements QuizModel {
     private static final Logger logger = LogsCenter.getLogger(QuizModelManager.class);
