+++ src/main/java/seedu/address/model/user/CardSrsData.java
@@ -29,7 +29,7 @@ public class CardSrsData {
     }
 
     /**
-     * Function for session managementmodel to get list of card datas
+     * Function for session management to get list of card datas
      * @param hashCode must be given
      * @return card from hashmap
      */
