+++ src/test/java/seedu/address/TestApp.java
@@ -9,8 +9,8 @@ import seedu.address.commons.core.Config;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.model.Lessons;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.modelmanager.managementmodel.ManagementModel;
-import seedu.address.model.modelmanager.managementmodel.ManagementModelManager;
+import seedu.address.model.modelmanager.management.ManagementModel;
+import seedu.address.model.modelmanager.management.ManagementModelManager;
 import seedu.address.storage.UserPrefsStorage;
 import seedu.address.testutil.TestUtil;
 
@@ -46,7 +46,7 @@ public class TestApp extends MainApp {
     }
 
     /**
-     * Returns a defensive copy of the managementmodel.
+     * Returns a defensive copy of the management.
      */
     public ManagementModel getModel() {
         ManagementModel copy = new ManagementModelManager(new UserPrefs(), new Lessons());
