+++ src/test/java/seedu/address/logic/commands/HistoryCommandTest.java
@@ -6,8 +6,8 @@ import org.junit.Test;
 
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.management.HistoryCommand;
-import seedu.address.model.modelmanager.managementmodel.ManagementModel;
-import seedu.address.model.modelmanager.managementmodel.ManagementModelManager;
+import seedu.address.model.modelmanager.management.ManagementModel;
+import seedu.address.model.modelmanager.management.ManagementModelManager;
 
 public class HistoryCommandTest {
     private CommandHistory history = new CommandHistory();
